+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -3,6 +3,7 @@
 * Added the `selenium-webdriver/testing/asserts` module. This module
     simplifies writing assertions against promised values (see
     example in module documentation).
+* Added the `webdriver.Capabilities` class.
 * Removed the deprecated `webdriver.Deferred#resolve` and
     `webdriver.promise.resolved` functions.